# Conflicts:
#	aaa-automation-tests/src/test/java/aaa/modules/e2e/templates/Scenario12.java
#	aaa-automation-tests/src/test/java/aaa/modules/e2e/templates/Scenario13.java
#	aaa-automation-tests/src/test/java/aaa/modules/regression/sales/auto_ss/functional/TestEValueNsfAHDRXX.java
#	aaa-automation-tests/src/test/resources/buglist.yaml
#	aaa-automation-tests/src/test/resources/testsuites/openl/auto_ca_suite.xml
#	aaa-automation-tests/src/test/resources/testsuites/openl/auto_ss_suite.xml
#	aaa-automation-tests/src/test/resources/testsuites/openl/openl.xml
#	aaa-automation-ui/src/main/java/aaa/main/enums/PolicyConstants.java
#	aaa-automation-ui/src/main/java/aaa/main/metadata/policy/AutoSSMetaData.java
#	aaa-automation-ui/src/main/java/aaa/main/modules/policy/auto_ca/defaulttabs/PremiumAndCoveragesTab.java
#	aaa-automation-ui/src/main/java/aaa/utils/excel/io/entity/area/ExcelCell.java
#	aaa-automation-ui/src/main/java/aaa/utils/excel/io/entity/area/table/ExcelTable.java

+++ aaa-automation-tests/src/test/java/aaa/helpers/openl/model/home_ca/HomeCaOpenLScheduledPropertyItem.java
@@ -7,7 +7,7 @@ public class HomeCaOpenLScheduledPropertyItem {
 	@ExcelTableColumnElement(name = OpenLFile.PRIMARY_KEY_COLUMN_NAME, isPrimaryKey = true)
 	private Integer number;
 
-	private Integer limit;
+	private Double limit;
 	private String propertyType;
 
 	public Integer getNumber() {
@@ -18,11 +18,11 @@ public class HomeCaOpenLScheduledPropertyItem {
 		this.number = number;
 	}
 
-	public Integer getLimit() {
+	public Double getLimit() {
 		return limit;
 	}
 
-	public void setLimit(Integer limit) {
+	public void setLimit(Double limit) {
 		this.limit = limit;
 	}
